🐞 fix: 自启延迟不能修改

+++ src/config.rs
@@ -52,14 +52,14 @@ pub struct AppConfig {
     #[serde(default = "bool_true")]
     pub show_github_button: bool,
     #[serde(default = "delay_default")]
-    pub on_boot_delay: i32,
+    pub on_boot_delay: f32,
 }
 
 #[rustfmt::skip]
 fn bool_true() -> bool { true }
 
 #[rustfmt::skip]
-fn delay_default() -> i32 { 3 }
+fn delay_default() -> f32 { 3.0 }
 
 pub fn on_boot_change(val: bool) {
     let auto = &AUTO_LAUNCH_CONF;
